Move event logging code from unit tests.
This PR moves the event logging classes (`Attribute`, `Event`,
`EventLogger`, etc) to a new header file so that we can reuse
them in multiple layers in the future.
